refactor: converted week dates store to typescript
