+++ store/week-dates/state.ts
@@ -1,0 +1,3 @@
+export default (): WeekDatesStoreState => ({
+  currentDate: new Date(),
+});
