+++ types/week-dates.d.ts
@@ -1,0 +1,3 @@
+interface WeekDatesStoreState {
+  currentDate: Date;
+}
